src/requirements.txt
asgiref==3.5.0
certifi==2022.12.7
Django==3.2.13
django-rest-framework==0.1.0
djangorestframework==3.13.1
Faker==13.4.0
python-dateutil==2.8.2
pytz==2022.1
schedule==1.1.0
sentry-sdk==1.12.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
urllib3==1.26.13